movie-main/movie-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
pandas
